Move express and mongoose config code to their own files + Update usages
